\name{print.foba}
\alias{print.foba}
\title{
Print a fitted foba model
}
\description{
print.foba() prints the variable selection path of a foba object.
}
\usage{
print.foba(x, ...)
}
\arguments{
\item{x}{
A fitted foba object
}
\item{...}{further arguments passed to or from other methods.}
}

\author{Tong Zhang}
\seealso{
predict.foba and foba
}

\keyword{methods}
\keyword{print}

